+++ extensions/quick-notes/src/components/actions.tsx
@@ -2,8 +2,8 @@ import { ActionPanel, Action, Icon, Clipboard, showToast, Toast } from "@raycast
 import CreateEditNoteForm from "./createEditNoteForm";
 import CreateTag from "./createTag";
 import DeleteNoteAction from "./deleteNoteAction";
-import { colors } from "../utils/utils";
-import { tagsAtom } from "../services/atoms";
+import { getSortHumanReadable, getTintColor } from "../utils/utils";
+import { menuAtom, sortArr, sortAtom, tagsAtom } from "../services/atoms";
 import { useAtom } from "jotai";
 import DeleteTags from "./deleteTags";
 
@@ -25,6 +25,8 @@ const Actions = ({
   createdAt?: Date;
 }) => {
   const [allTags] = useAtom(tagsAtom);
+  const [sort, setSort] = useAtom(sortAtom);
+  const [_, setMenu] = useAtom(menuAtom);
   return (
     <ActionPanel>
       <ActionPanel.Section>
@@ -34,7 +36,7 @@ const Actions = ({
               title="Edit Note"
               icon={{
                 source: Icon.Pencil,
-                tintColor: colors.find((c) => c.name === "sky")?.tintColor,
+                tintColor: getTintColor("sky"),
               }}
               target={
                 <CreateEditNoteForm isDraft={isDraft} title={title} note={note} tags={tags} createdAt={createdAt} />
@@ -42,7 +44,7 @@ const Actions = ({
             />
             <Action
               title="Copy Note"
-              icon={{ source: Icon.CopyClipboard, tintColor: colors.find((c) => c.name === "turquoise")?.tintColor }}
+              icon={{ source: Icon.CopyClipboard, tintColor: getTintColor("turquoise") }}
               shortcut={{ modifiers: ["cmd", "shift"], key: "c" }}
               onAction={() => {
                 Clipboard.copy(note ?? "").then(() => {
@@ -56,7 +58,7 @@ const Actions = ({
           title="New Note"
           icon={{
             source: Icon.PlusSquare,
-            tintColor: colors.find((c) => c.name === "green")?.tintColor,
+            tintColor: getTintColor("green"),
           }}
           target={<CreateEditNoteForm isDraft={true} />}
           shortcut={{ modifiers: ["cmd"], key: "n" }}
@@ -68,7 +70,7 @@ const Actions = ({
           title="Filter Tag"
           icon={{
             source: Icon.Filter,
-            tintColor: colors.find((c) => c.name === "turquoise")?.tintColor,
+            tintColor: getTintColor("turquoise"),
           }}
           shortcut={{ modifiers: ["cmd", "shift"], key: "t" }}
         >
@@ -79,7 +81,7 @@ const Actions = ({
                 key={i}
                 icon={{
                   source: "dot.png",
-                  tintColor: colors.find((c) => c.name === tag.color)?.tintColor ?? "blue",
+                  tintColor: getTintColor(tag.color) ?? "blue",
                 }}
                 title={tag.name}
                 onAction={() => {
@@ -93,7 +95,7 @@ const Actions = ({
           title="New Tag"
           icon={{
             source: Icon.Tag,
-            tintColor: colors.find((c) => c.name === "turquoise")?.tintColor,
+            tintColor: getTintColor("turquoise"),
           }}
           target={<CreateTag />}
           shortcut={{ modifiers: ["cmd"], key: "t" }}
@@ -102,11 +104,47 @@ const Actions = ({
           title="Delete Tags"
           icon={{
             source: Icon.Trash,
-            tintColor: colors.find((c) => c.name === "red")?.tintColor,
+            tintColor: getTintColor("red"),
           }}
           target={<DeleteTags />}
           shortcut={{ modifiers: ["ctrl", "shift"], key: "t" }}
         />
+      </ActionPanel.Section>
+      <ActionPanel.Section>
+        <Action
+          title="Toggle Menu"
+          icon={{
+            source: Icon.AppWindowSidebarRight,
+            tintColor: getTintColor("indigo"),
+          }}
+          onAction={() => setMenu()}
+          shortcut={{ modifiers: ["cmd"], key: "m" }}
+        />
+        <ActionPanel.Submenu
+          title="Sort"
+          icon={{
+            source: Icon.Filter,
+            tintColor: getTintColor("indigo"),
+          }}
+          shortcut={{ modifiers: ["cmd"], key: "s" }}
+        >
+          {sortArr.map((s, i) => (
+            <Action
+              key={i}
+              icon={
+                s === sort
+                  ? {
+                      source: Icon.ArrowRightCircle,
+                      tintColor: "teal",
+                    }
+                  : undefined
+              }
+              title={getSortHumanReadable(s)}
+              onAction={() => setSort(s)}
+            />
+          ))}
+          <Action.Push title="Create" icon={Icon.Plus} target={<CreateTag />} />
+        </ActionPanel.Submenu>
       </ActionPanel.Section>
     </ActionPanel>
   );