+++ extensions/quick-notes/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Quick Notes Changelog
 
+## [Info Menu] - 2024-05-23
+
+- New Info Menu to show details and tags for a note using `⌘ + M`
+- Sort straight from the notes list instead of preferences using `⌘ + S`'
+- Both these new features save your settings in LocalStorage
+
 ## [Copy Action] - 2024-05-08
 
 - New `Copy to Clipboard` action on note body using `⌘ + ⇧ + C`